Move utility functions to manager.utils:
* jira_ticket, jira_comment, get_jira_title, get_jira_description

* upload_to_lfa
