+++ manager/api/views.py
@@ -50,7 +50,7 @@ from manager.settings import (
     AUTH_LDAP_2_SERVER_URI,
     AUTH_LDAP_3_SERVER_URI,
 )
-from manager.utils import CommandPermission
+from manager.utils import CommandPermission, get_jira_title, get_jira_description, jira_ticket, jira_comment, upload_to_lfa
 
 valid_response = openapi.Response("Valid token", TokenSerializer)
 invalid_response = openapi.Response("Invalid token")
@@ -892,7 +892,6 @@ def tcs_main_command(request, *args, **kwargs):
     response = requests.post(url, json=request.data)
     return Response(response.json(), status=response.status_code)
 
-
 @api_view(["GET"])
 @permission_classes((IsAuthenticated,))
 def tcs_main_docstrings(request, *args, **kwargs):
@@ -915,51 +914,6 @@ def tcs_main_docstrings(request, *args, **kwargs):
     url = f"http://{os.environ.get('COMMANDER_HOSTNAME')}:{os.environ.get('COMMANDER_PORT')}/tcs/main/docstrings"
     response = requests.get(url)
     return Response(response.json(), status=response.status_code)
-
-
-# @api_view(["POST"])
-# @permission_classes((IsAuthenticated,))
-def lfa(request, *args, **kwargs):
-    """Connects to LFA API to upload a new file
-
-    Params
-    ------
-    request: Request
-        The Request object
-    args: list
-        List of addittional arguments. Currently unused
-    kwargs: dict
-        Dictionary with request arguments. Currently unused
-
-    Returns
-    -------
-    Response
-        The response and status code of the request to the LOVE-commander LFA API
-    """
-
-    option = kwargs.get("option", None)
-    url = f"http://{os.environ.get('COMMANDER_HOSTNAME')}:{os.environ.get('COMMANDER_PORT')}/lfa/{option}"
-
-    if len(request.FILES.getlist("file[]")) == 0:
-        return Response({"ack": "No files to upload"}, status=400)
-
-    if option == "upload-file":
-        uploaded_files_urls = []
-        files_to_upload = request.FILES.getlist("file[]")
-        for file in files_to_upload:
-            upload_file_response = requests.post(url, files={"uploaded_file": file})
-            if upload_file_response.status_code == 200:
-                uploaded_files_urls.append(upload_file_response.json().get("url"))
-
-        if len(uploaded_files_urls) != len(files_to_upload):
-            return Response({"ack": "Error when uploading files"}, status=400)
-
-        return Response(
-            {"ack": "All files uploaded correctly", "urls": uploaded_files_urls},
-            status=200,
-        )
-
-    return Response({"ack": "Option not found"}, status=400)
 
 
 class CSCAuthorizationRequestViewSet(
@@ -1146,275 +1100,6 @@ class CSCAuthorizationRequestViewSet(
         return Response({"error": "Bad request"}, status=status.HTTP_400_BAD_REQUEST)
 
 
-def get_jira_title(request_data):
-    """Generate title for Jira ticket
-
-    Parameters:
-    -----------
-    request_data: dict
-        Request data
-
-    Returns:
-    --------
-    title: str
-        Jira ticket title
-    """
-    request_type = request_data.get("request_type")
-    jira_title = request_data.get("jira_issue_title")
-
-    if jira_title is not None and jira_title != "":
-        return jira_title
-
-    return "LOVE generated: " + request_type
-
-
-def get_jira_description(request_data):
-    """Generate description for Jira ticket
-
-    Parameters:
-    -----------
-    request_data: dict
-        Request data
-
-    Returns:
-    --------
-    description: str
-        Jira ticket description
-
-    Raises:
-    -------
-    Exception
-        If there is an error reading the request data
-    """
-    # Shared params
-    request_type = request_data["request_type"]
-    try:
-        lfa_files_urls = request_data["lfa_files_urls"]
-        message_log = request_data["message_text"]
-        user_id = request_data["user_id"]
-        user_agent = request_data["user_agent"]
-    except Exception as e:
-        raise Exception("Error reading params") from e
-
-    # Exposure log params
-    if request_type == "exposure":
-        try:
-            obs_id = request_data["obs_id"]
-            instrument = request_data["instrument"]
-            exposure_flag = request_data["exposure_flag"]
-        except Exception as e:
-            raise Exception("Error reading params") from e
-        description = (
-            "*Created by* "
-            + user_id
-            + " *from* "
-            + user_agent
-            + "\n"
-            + "*Observation ids:* "
-            + str(obs_id)
-            + "\n"
-            + "*Instrument:* "
-            + instrument
-            + "\n"
-            + "*Exposure flag:* "
-            + exposure_flag
-            + "\n"
-            + "*Files:* "
-            + "\n"
-            + str(lfa_files_urls)
-            + "\n\n"
-            + message_log
-        )
-    # Narrative log params
-    if request_type == "narrative":
-        try:
-            systems = (
-                ", ".join(request_data["systems"].split(","))
-                if request_data.get("systems", False)
-                else "None"
-            )
-            subsystems = (
-                ", ".join(request_data["subsystems"].split(","))
-                if request_data.get("subsystems", False)
-                else "None"
-            )
-            cscs = (
-                ", ".join(request_data["cscs"].split(","))
-                if request_data.get("cscs", False)
-                else "None"
-            )
-            begin_date = request_data["date_begin"]
-            end_date = request_data["date_end"]
-            time_lost = str(request_data["time_lost"])
-        except Exception as e:
-            raise Exception("Error reading params") from e
-
-        description = (
-            "*Created by* "
-            + user_id
-            + " *from* "
-            + user_agent
-            + "\n"
-            + "*Time of incident:* "
-            + begin_date
-            + " *-* "
-            + end_date
-            + "\n"
-            + "*Time lost:* "
-            + time_lost
-            + "\n"
-            + "*System:* "
-            + systems
-            + "\n"
-            + "*Subsystems:* "
-            + subsystems
-            + "\n"
-            + "*CSCs:* "
-            + cscs
-            + "\n"
-            + "*Files:* "
-            + "\n"
-            + str(lfa_files_urls)
-            + "\n\n"
-            + message_log
-        )
-
-    return description if description is not None else ""
-
-
-def jira_ticket(request):
-    """Connects to JIRA API to create a ticket on a specific project.
-    For more information on issuetypes refer to:
-    ttps://jira.lsstcorp.org/rest/api/latest/issuetype/?projectId=JIRA_PROJECT_ID
-
-    Params
-    ------
-    request: Request
-        The request object from the Django view
-
-    Returns
-    -------
-    Response
-        The response and status code of the request to the JIRA API
-
-    Raises
-    ------
-    Exception
-        If there is an error reading the request data
-    """
-    request_data = request.data
-    if "request_type" not in request_data:
-        return Response({"ack": "Error reading request type"}, status=400)
-
-    tags_data = (
-        request_data.get("tags").split(",")
-        if request_data.get("tags") != "undefined"
-        else []
-    )
-
-    try:
-        jira_payload = {
-            "fields": {
-                "project": {"id": os.environ.get("JIRA_PROJECT_ID")},
-                "labels": [
-                    "LOVE",
-                    *tags_data,
-                ],
-                "summary": get_jira_title(request_data),
-                "description": get_jira_description(request_data),
-                "customfield_15602": "on"
-                if int(request_data.get("level", 0)) >= 100
-                else "off",  # Is Urgent?
-                "customfield_16702": float(
-                    request_data.get("time_lost", 0)
-                ),  # Obs. time loss
-                "issuetype": {"id": 12302},
-            },
-            "update": {"components": [{"set": [{"name": "LOVE"}]}]},
-        }
-    except Exception as e:
-        return Response({"ack": f"Error creating jira payload: {e}"}, status=400)
-
-    headers = {
-        "Authorization": f"Basic {os.environ.get('JIRA_API_TOKEN')}",
-        "content-type": "application/json",
-    }
-    url = f"https://{os.environ.get('JIRA_API_HOSTNAME')}/rest/api/latest/issue/"
-    response = requests.post(url, json=jira_payload, headers=headers)
-    response_data = response.json()
-    if response.status_code == 201:
-        return Response(
-            {
-                "ack": "Jira ticket created",
-                "url": f"https://jira.lsstcorp.org/browse/{response_data['key']}",
-            },
-            status=200,
-        )
-
-    return Response(
-        {
-            "ack": "Jira ticket could not be created",
-            "error": response_data,
-        },
-        status=400,
-    )
-
-
-def jira_comment(request):
-    """Connects to JIRA API to add a comment to a previously created ticket on a specific project.
-    For more information on issuetypes refer to:
-    https://jira.lsstcorp.org/rest/api/latest/issuetype/?projectId=JIRA_PROJECT_ID
-
-    Params
-    ------
-    request: Request
-        The Request object
-
-    Returns
-    -------
-    Response
-        The response and status code of the request to the JIRA API
-
-    Raises
-    ------
-    Exception
-        If there is an error reading the request data
-    """
-    request_data = request.data
-    if "jira_issue_id" not in request_data:
-        return Response({"ack": "Error reading the JIRA issue ID"}, status=400)
-
-    jira_id = request_data.get("jira_issue_id")
-
-    try:
-        jira_payload = {
-            "body": get_jira_description(request_data),
-        }
-    except Exception as e:
-        return Response({"ack": f"Error creating jira payload: {e}"}, status=400)
-
-    headers = {
-        "Authorization": f"Basic {os.environ.get('JIRA_API_TOKEN')}",
-        "content-type": "application/json",
-    }
-    url = f"https://{os.environ.get('JIRA_API_HOSTNAME')}/rest/api/latest/issue/{jira_id}/comment"
-    response = requests.post(url, json=jira_payload, headers=headers)
-    if response.status_code == 201:
-        return Response(
-            {
-                "ack": "Jira comment created",
-                "url": f"https://jira.lsstcorp.org/browse/{jira_id}",
-            },
-            status=200,
-        )
-    return Response(
-        {
-            "ack": "Jira comment could not be created",
-        },
-        status=400,
-    )
-
-
 @swagger_auto_schema(
     method="get",
     responses={
@@ -1511,7 +1196,7 @@ class ExposurelogViewSet(viewsets.ViewSet):
         lfa_urls = []
         files_to_upload = request.FILES.getlist("file[]")
         if len(files_to_upload) > 0:
-            lfa_response = lfa(request, option="upload-file")
+            lfa_response = upload_to_lfa(request, option="upload-file")
             if lfa_response.status_code != 200:
                 return lfa_response
             lfa_urls = lfa_response.data.get("urls")
@@ -1527,9 +1212,9 @@ class ExposurelogViewSet(viewsets.ViewSet):
 
             jira_response = None
             if request.data.get("jira_new") == "true":
-                jira_response = jira_ticket(request)
+                jira_response = jira_ticket(request.data)
             else:
-                jira_response = jira_comment(request)
+                jira_response = jira_comment(request.data)
 
             if jira_response.status_code == 400:
                 return Response(
@@ -1545,6 +1230,7 @@ class ExposurelogViewSet(viewsets.ViewSet):
         if "file[]" in json_data:
             del json_data["file[]"]
 
+        # Split lists of values separated by comma
         array_keys = {"tags"}
         for key in array_keys:
             if key in json_data:
@@ -1579,7 +1265,7 @@ class ExposurelogViewSet(viewsets.ViewSet):
         lfa_urls = []
         files_to_upload = request.FILES.getlist("file[]")
         if len(files_to_upload) > 0:
-            lfa_response = lfa(request, option="upload-file")
+            lfa_response = upload_to_lfa(request, option="upload-file")
             if lfa_response.status_code != 200:
                 return lfa_response
             lfa_urls = lfa_response.data.get("urls")
@@ -1592,7 +1278,6 @@ class ExposurelogViewSet(viewsets.ViewSet):
             del json_data["file[]"]
 
         array_keys = {
-            "urls",
             "tags",
         }
         for key in array_keys:
@@ -1652,7 +1337,7 @@ class NarrativelogViewSet(viewsets.ViewSet):
         lfa_urls = []
         files_to_upload = request.FILES.getlist("file[]")
         if len(files_to_upload) > 0:
-            lfa_response = lfa(request, option="upload-file")
+            lfa_response = upload_to_lfa(request, option="upload-file")
             if lfa_response.status_code != 200:
                 return lfa_response
             lfa_urls = lfa_response.data.get("urls")
@@ -1668,9 +1353,9 @@ class NarrativelogViewSet(viewsets.ViewSet):
 
             jira_response = None
             if request.data.get("jira_new") == "true":
-                jira_response = jira_ticket(request)
+                jira_response = jira_ticket(request.data)
             else:
-                jira_response = jira_comment(request)
+                jira_response = jira_comment(request.data)
 
             if jira_response.status_code == 400:
                 return Response(
@@ -1686,11 +1371,8 @@ class NarrativelogViewSet(viewsets.ViewSet):
         if "file[]" in json_data:
             del json_data["file[]"]
 
+        # Split lists of values separated by comma
         array_keys = {
-            "tags",
-            "systems",
-            "subsystems",
-            "cscs",
             "components",
             "primary_software_components",
             "primary_hardware_components",
@@ -1724,7 +1406,7 @@ class NarrativelogViewSet(viewsets.ViewSet):
         lfa_urls = []
         files_to_upload = request.FILES.getlist("file[]")
         if len(files_to_upload) > 0:
-            lfa_response = lfa(request, option="upload-file")
+            lfa_response = upload_to_lfa(request, option="upload-file")
             if lfa_response.status_code != 200:
                 return lfa_response
             lfa_urls = lfa_response.data.get("urls")
@@ -1737,11 +1419,6 @@ class NarrativelogViewSet(viewsets.ViewSet):
             del json_data["file[]"]
 
         array_keys = {
-            "urls",
-            "tags",
-            "systems",
-            "subsystems",
-            "cscs",
             "components",
             "primary_software_components",
             "primary_hardware_components",
